+++ include/cucumber-cpp/defs.hpp
@@ -1,6 +1,5 @@
-#include "internal/step/StepManager.hpp"
-#include "internal/hook/HookRegistrar.hpp"
-#include "internal/ContextManager.hpp"
-
-#include "internal/Macros.hpp"
-#include "internal/drivers/DriverSelector.hpp"
+#warning Use of defs.hpp is deprecated, please use either autodetect.hpp or generic.hpp
+#include "internal/defs.hpp"
+#ifndef STEP_INHERITANCE
+    #include "internal/drivers/GenericDriver.hpp"
+#endif