+++ include/cucumber-cpp/internal/drivers/DriverSelector.hpp
@@ -1,9 +1,7 @@
 #if defined(GTEST_INCLUDE_GTEST_GTEST_H_)
-#include "GTestDriver.hpp"
+    #include "GTestDriver.hpp"
 #elif defined(BOOST_TEST_CASE)
-#include "BoostDriver.hpp"
+    #include "BoostDriver.hpp"
 #elif defined(CPPSPEC_H_)
-#include "CppSpecDriver.hpp"
-#else // No test framework
-#include "FakeDriver.hpp"
+    #include "CppSpecDriver.hpp"
 #endif