+++ tests/integration/drivers/GTestDriverTest.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include <cucumber-cpp/defs.hpp>
+#include <cucumber-cpp/autodetect.hpp>
 
 #include "../../utils/DriverTestRunner.hpp"