+++ tests/utils/CukeCommandsFixture.hpp
@@ -2,7 +2,7 @@
 #define CUKE_CUKECOMMANDSFIXTURE_HPP_
 
 #include <cucumber-cpp/internal/CukeCommands.hpp>
-#include <cucumber-cpp/internal/drivers/FakeDriver.hpp>
+#include <cucumber-cpp/internal/drivers/GenericDriver.hpp>
 #include "StepManagerTestDouble.hpp"
 
 #include <boost/shared_ptr.hpp>
@@ -10,7 +10,7 @@
 using namespace cucumber::internal;
 using boost::shared_ptr;
 
-class EmptyStep : public FakeStep {
+class EmptyStep : public GenericStep {
     void body() {}
 };